Created Basic Rocket Script
Created new projectile type that interacts with surfaces and has placeholder method for creating explosion at its location upon surface collision. Rocket takes the mouse location as an input to then calculate the horizontal and vertical movement needed to go at the same angle as the mouse is from the spawn location (player), calculating the ratio of the two right triangles via the hypotenuse of the two triangles.