Gameplay Experience Adjustments
A lot of adjustments regarding movement, fire rate, jumping, gravity, and knockback. These series of changes are primarily for making the movement values more realistic for the environment and level, as well as making the difficulty of the game more about the aiming of your jumps rather than a result of clunky player controls.